Various fixes.
Install the openssh-clients package, as well as several IPMI related
items.

Set the root password (to 'centos' in this example) so that sshd works
without a problem. It is recommended to change this password for production
use obviously :)

Make the network start automatically as well, with eth0 being configured
for DHCP. This should be configurable in the future.

Fixes #1